Replace the session list with a tree view
The approach I took was to automatically create a hierarchy tree by breaking
saved session names on slashes ('/'), but only for viewing in the config window.
Sessions are still saved flatly in the registry, and if you use a PuTTY version
that uses a session list you'll just see a long list of sessions with slashes in
their names. This keeps previous versions from breaking.
I needed to add dialog.h & dialog.c to the puttygen project since the C file
contains the implementation of the tree control internal management code. This
should not cause any platform dependency issues.

This was a cherry-pick from git://github.com/leeor/putty.git / master branch..
commit 43a24203de6c4f20693b60d7089408fdfbfc4b3f