GH-89: drop /safeseh so we can build on x64 again
This flag shouldn't really be necessary; it appears to just be an assert
that your exception code is sane.  We're not actually using any languages
that have exceptions, and hopefully any Windows code that wants to use
exceptions within us is safe.

Maybe it's worth turning this on for x32 runs of mkfiles, just to check?