Fixing a bug in fetching session list from files
While fetching session list from files, ignore folders and fetch next
available session file instead of returning null.

If you happen to have a stray folder in the sessions folder, then the
order in which FindNextFile() gets files is uncertain. And often we end up
seeing a folder traversed in the middle of several session files.

This issue is easily reproducable if we have lots of session files that
start with different names and a folder.